% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/displayOptPath.R
\name{displayOptPath}
\alias{displayOptPath}
\title{Display MOGSA's Optimization Path as a HTML or LaTeX table.}
\usage{
displayOptPath(opt.path, format = "html", ...)
}
\arguments{
\item{opt.path}{[\code{\link{data.frame}}]\cr
Optimization path returned by \code{\link{runMOGSA}}.}

\item{format}{[\code{\link{character}(1L)}]\cr
Type of table being generated; either \code{"html"} (default) or \code{"latex"}.}

\item{...}{[any]\cr
Further arguments to be passed to \code{kable}.}
}
\value{
Returns MOGSA's optimization path either as HTML (default) or LaTeX table.
}
\description{
Convenience function for displaying the optimization path of MOGSA as
  a prettified (HTML or LaTeX) table.
}
\examples{
## Example 1:
# (i) Define two single-objective test problems (fn1, fn2) and a starting point (ind):
fn1 = function(x) sum((x - c(2, 0))^2)
fn2 = function(x) sum((x - c(0, 1))^2)
ind = c(1, 1)

# (ii) Run MOGSA:
mogsa.result = runMOGSA(ind = ind, fn1 = fn1, fn2 = fn2)

# (iii) Display the optimization path of MOGSA as HTML (or LaTeX) table:
displayOptPath(mogsa.result)
}
