% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{class}
\name{mgsa-class}
\alias{mgsa-class}
\title{Class \code{"mgsa"}}
\description{
mgsa class here.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("mgsa", ...)}.
}

\examples{

  showClass("mgsa")
  # library(mogsa)
  # loading gene expression data and supplementary data
  data(NCI60_4array_supdata)
  data(NCI60_4arrays)
  # split gene set annotation into two sets.
  sup1 <- lapply(NCI60_4array_supdata, function(x) x[, 1:10])
  sup2 <- lapply(NCI60_4array_supdata, function(x) x[, -(1:10)])
  # project two sets of annotation
  mgsa1 <- mogsa(x = NCI60_4arrays, sup=sup1, nf=9,
                proc.row = "center_ssq1", w.data = "inertia", statis = TRUE)
  mgsa2 <- mogsa(x = NCI60_4arrays, sup=sup2, nf=9,
                 proc.row = "center_ssq1", w.data = "inertia", statis = TRUE)
  # combine two indenpendent mgsa sets
  mgsa_comb <- combine(mgsa1, mgsa2)
  dim(getmgsa(mgsa1, "fac.scr"))
  dim(getmgsa(mgsa2, "fac.scr"))
  dim(getmgsa(mgsa_comb, "fac.scr"))

}
\seealso{
\code{\linkS4class{moa}} and \code{\linkS4class{moa.sup}}
}
\author{
Chen Meng
}
\keyword{classes}
