% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{class}
\name{moa-class}
\alias{moa-class}
\alias{plot,moa,missing-method}
\title{Class \code{"moa"}}
\description{
moa class object
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("moa", ...)}. %% ~~ describe objects here ~~
}

\examples{

    showClass("moa")
    # load("R/mogsa/data/NCI60_4arrays.rda")
    data(NCI60_4arrays)
    ana <- moa(NCI60_4arrays, proc.row = "center_ssq1", w.data = "inertia", statis = TRUE)

    plot(ana, value="eig")
    plot(ana, value="tau", type=2)

}
\references{
Herve Abdi, Lynne J. Williams, Domininique Valentin and Mohammed
Bennani-Dosse. STATIS and DISTATIS: optimum multitable principal component
analysis and three way metric multidimensional scaling. WIREs Comput Stat
2012. Volume 4, Issue 2, pages 124-167

Herve Abdi, Lynne J. Williams, Domininique Valentin. Multiple factor
analysis: principal component analysis for multitable and multiblock data
sets. WIREs Comput Stat 2013
}
\author{
Chen Meng
}
\keyword{classes}
