% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{class}
\name{moa.sup-class}
\alias{moa.sup-class}
\title{Class \code{"moa.sup"}}
\description{
moa.sup class desc.
}
\section{Objects from the Class}{
 Objects can be created by calls of the form
\code{new("moa.sup", ...)}.
}

\examples{

	showClass("moa.sup")
	data(NCI60_4array_supdata)
	data(NCI60_4arrays)

	sapply(NCI60_4array_supdata, dim)
	ana <- moa(NCI60_4arrays, proc.row = "center_ssq1", w.data = "inertia", statis = TRUE)
	plot(ana, value="eig")
	smoa <- sup.moa(ana, sup=NCI60_4array_supdata, nf=5)

}
\seealso{
objects to See Also as \code{\link{decompose.gs.ind}},
\code{\link{box.gs.feature}}, \code{\link{plotGS}},
\code{\link{decompose.gs.group}}.
}
\author{
Chen Meng
}
\keyword{classes}
