% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/moaExtract.R
\name{moaScore}
\alias{moaScore}
\title{Extract global scores from an object of class \code{\link{moa-class}}.}
\usage{
moaScore(moa)
}
\arguments{
\item{moa}{An object of class \code{\link{moa-class}}}
}
\value{
A matrix of global score
}
\description{
Extract global scores from an object of class \code{\link{moa-class}}.
}
\examples{

# see examples in \\code{\\link{mbpca}}

data("NCI60_4arrays")
moa <- mbpca(NCI60_4arrays, ncomp = 10, k = "all", method = "globalScore", option = "lambda1", 
             center=TRUE, scale=FALSE)

genes <- moaCoef(moa)
scr <- moaScore(moa)


}
\seealso{
\code{\link{moaCoef}}
}
\author{
Chen Meng
}
