% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGeneSet.R
\name{prepGraphite}
\alias{prepGraphite}
\title{Prepare pathway gene sets from graphite package}
\usage{
prepGraphite(db, id = c("entrez", "symbol"))
}
\arguments{
\item{db}{The database to be used, an object of class either 'PathwayList'
create by "pathways" function.}

\item{id}{Which identifier for output, either "entrez" or "symbol".}
}
\value{
This function returns an object of list containing gene set
information, which could be further processed by function "prepSupMoa" to
convert to the object that can be used as input of "sup.moa" or "mogsa".
}
\description{
Prepare pathway gene sets from "graphite" package, which could be passed to
"prepSupMoa" function.
}
\details{
Only support "entrez" or "symbol" output currently.
}
\examples{

  library(graphite)
  keggdb <- prepGraphite(db = pathways("hsapiens", "kegg")[1:3], id = "entrez")

}
\references{
Sales G, Calura E and Romualdi C (2014). graphite: GRAPH
Interaction from pathway Topological Environment. R package version 1.10.1.
}
\seealso{
See Also as \code{\link{prepMsigDB}} and \code{\link{prepSupMoa}}.
}
\author{
Chen Meng
}
\keyword{graphite}
\keyword{pahtways}
