% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preGeneSet.R
\name{prepMsigDB}
\alias{prepMsigDB}
\title{Conver gmt format file to a list}
\usage{
prepMsigDB(file)
}
\arguments{
\item{file}{The directory and file name of the gmt file.}
}
\value{
This function returns an object of list containing gene set
information, which could be further processed by function "prepSupMoa" to
convert to the object that can be used as input of "sup.moa" or "mogsa".
}
\description{
Convert a gmt file (Could be downloaded from MSigDB) to a list of gene sets
information.
}
\examples{

	# not run
	dir <- system.file(package = "mogsa")
	preGS <- prepMsigDB(file=paste(dir, 
		"/extdata/example_msigdb_data.gmt.gz", sep = ""))

}
\seealso{
See Also as \code{\link{prepGraphite}} and
\code{\link{prepSupMoa}}.
}
\author{
Chen Meng
}
