% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{methods}
\name{print-methods}
\alias{print-methods}
\alias{print}
\alias{print,moa-method}
\alias{print,moa.sup-method}
\alias{print,mgsa-method}
\title{Methods for function \code{print}}
\description{
Methods for function \code{print}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"moa\")")}{ print "moa" class }

\item{list("signature(object = \"moa.sup\")")}{ print "sup.moa" class }

\item{list("signature(object = \"mgsa\")")}{ print "mgsa" class } }
}

\keyword{methods}
\keyword{mgsa-class}
\keyword{moa-class}
\keyword{sup.moa-class}
