% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{methods}
\name{show-methods}
\alias{show-methods}
\alias{show}
\alias{show,moa-method}
\alias{show,moa.sup-method}
\alias{show,mgsa-method}
\title{Methods for function \code{show}}
\description{
Methods for function \code{show}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"moa\")")}{ show "moa" class }

\item{list("signature(object = \"moa.sup\")")}{ show "sup.moa" class }

\item{list("signature(object = \"mgsa\")")}{ show "mgsa" class } }
}

\keyword{methods}
\keyword{mgsa-class}
\keyword{moa-class}
\keyword{sup.moa-class}
