% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mogsa-package.R
\docType{methods}
\name{summary-methods}
\alias{summary-methods}
\alias{summary}
\alias{summary,moa-method}
\alias{summary,moa.sup-method}
\alias{summary,mgsa-method}
\title{Methods for function \code{summary}}
\description{
Methods for function \code{summary}
}
\section{Methods}{
 \describe{

\item{list("signature(object = \"moa\")")}{ summary "moa" class }

\item{list("signature(object = \"moa.sup\")")}{ summary "sup.moa" class }

\item{list("signature(object = \"mgsa\")")}{ summary "mgsa" class } }
}

\keyword{methods}
\keyword{mgsa-class}
\keyword{moa-class}
\keyword{sup.moa-class}
