% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.nipals.R
\name{svd.nipals}
\alias{svd.nipals}
\title{singular valude decomposition using NIPALS algorithm}
\usage{
svd.nipals(x, nf, ...)
}
\arguments{
\item{x}{a numeric matrix whose SVD decomposition is to be computed}

\item{nf}{the number of singular vectors want to computed}

\item{...}{other parameters passed to \code{\link[nipals]{nipals}}, except
the following argument:
center (=FALSE), scale (=FALSE), gramschmidt (=FALSE)#'}
}
\value{
the same as svd, list of three components, d, u, v
}
\description{
The function may be faster when x is a sparseMatrix object and 
  a few components need to be calculated. In addition, NA values could be 
  presented in x.
}
\examples{
# small matrix
m <- matrix(rnorm(200), 20, 10)
s <- svd.niplas(m)

# large sparse matrix
m <- matrix(sample(c(rnorm(10000), rep(0, 9990000))), 10000, 1000)
ms <- Matrix(m, sparse=TRUE)
system.time(
  s1 <- svd(m)
)
system.time(
  s2 <- svd.niplas(ms, nf = 1)
) 

}
\seealso{
\code{\link{svd}}
}
