% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/svd.nipals.R
\name{svd.solver}
\alias{svd.solver}
\title{Singular Value Decomposition of a Matrix}
\usage{
svd.solver(x, nf, opts.svd.nipals = list(), opts.svds = list(),
  opts.fast.svd = list(), opts.svd = list())
}
\arguments{
\item{x}{a numeric or complex matrix whose SVD decomposition is to be computed.}

\item{nf}{the number of singular vectors want to computed. This will be ignored 
if fast.svd or svd function is used.}

\item{opts.svd.nipals}{A list of named parameters passed to \code{\link{svd.nipals}}}

\item{opts.svds}{A list of named parameters passed to \code{\link[RSpectra]{svds}}}

\item{opts.svd.fast.svd}{A list of named parameters passed to \code{\link[corpcor]{fast.svd}}}

\item{opts.svd.svd}{A list of named parameters passed to \code{\link[base]{svd}}}
}
\description{
different methods are automatically selected to calculate svd
  according to matrix properties, such as missing values, object class, size.
}
\details{
There are 4 different options could be used:
  1. \code{\link{svd.nipals}} When there are missing values in the matrix
  2. \code{\link[RSpectra]{svds}} when x is an object of class \code{Matrix} or only a small 
    number of singular vectors to be calculated (<3)
  3. \code{\link[corpcor]{fast.svd}} when x is big fat or thin matrix
  4. \code{\link[base]{svd}} if not any other cases
}
\examples{
m <- matrix(sample(c(rnorm(1000), rep(0, 999000))), 10000, 100)
decomp <- svd.solver(m)
attr(decomp, "solver")

decomp <- svd.solver(m[1:200, ])
attr(decomp, "solver")

ms <- Matrix(m, sparse=TRUE)
decomp <- svd.solver(ms, nf = 2)
attr(decomp, "solver")

mm <- Matrix(m)
decomp <- svd.solver(mm, nf = 2)
attr(decomp, "solver")

mna <- m
mna[sample(1:length(mna), size = 100)] <- NA
decomp <- svd.solver(mna, nf = 2)
attr(decomp, "solver")

mnas <- Matrix(mna, sparse = TRUE)
decomp <- svd.solver(mnas, nf = 2)
attr(decomp, "solver")
}
\author{
Chen Meng
}
