% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toMoa.R
\name{toMoa}
\alias{toMoa}
\title{convert \code{\link{mbpca}} result to \code{\link{moa-class}}}
\usage{
toMoa(data, x, call)
}
\arguments{
\item{data}{The preprocessed data in \code{\link{mbpca}}}

\item{x}{The object calculated in \code{\link{mbpca}}}

\item{call}{The call of mbpca}
}
\value{
An object of \code{moa-class}.
}
\description{
An internal function called by \code{\link{mbpca}}.
}
\author{
Chen Meng
}
